fix: apply all 8 review thread comments from round 3
Agent-Logs-Url: https://github.com/schivei/aeternum-db/sessions/15240e01-863e-4e8e-a35d-b2f73f2c0de4

+++ core/src/query/cost_model.rs
@@ -92,7 +92,7 @@ impl CostModel {
     ///
     /// Complexity is O(left × right).
     pub fn estimate_nested_loop_cost(&self, left_rows: usize, right_rows: usize) -> f64 {
-        (left_rows * right_rows) as f64 * self.cpu_cost_factor
+        (left_rows as f64) * (right_rows as f64) * self.cpu_cost_factor
     }
 
     /// Estimate the cost of a hash join.
